third_party/blink/web_tests/fast/forms/select/listbox-tap-expected.txt

Tapping on listbox items should toggle selection.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getSelectedValues(select) is ""
PASS getSelectedValues(select) is "1"
PASS getSelectedValues(select) is "1,2"
PASS getSelectedValues(select) is "1,2,3"
PASS getSelectedValues(select) is "1,2,3"
PASS getSelectedValues(select) is "1,3"
PASS getSelectedValues(select) is "3"
PASS getSelectedValues(select) is ""
PASS successfullyParsed is true

TEST COMPLETE

1
2
3
